Reorganizando a estrutura do trabalho
Criei arquivos de handlers para os files e os eventos criei um arquivo
types.h que contem nossas definicoes de tipos e constantes, e modifiquei
o main completamente.

+++ TrabalhoCompGraf_1/bspline.h
@@ -19,7 +19,7 @@ private:
 	vector<float> m_knots;
 
 	float basis(int, int, float);
-	void createKnotsVector(int, int, int);
+	void createKnotsVector(int, int, size_t);
 
 };
